AITeacher
Computer-Assisted Instruction (Project Name: AITeacher) - The use of computers in education is referred to as computer-assisted instruction (CAI). This program will help an elementary-school student learn multiplication. I used a Random object to produce two positive one-digit integers. The program should then prompt the user with a question, such as How much is 6 times 7? The student then inputs the answer. Next, the program checks the student’s answer. If it is correct, display the message "Very good!" and ask another multiplication question. If the answer is wrong, display the message "No. Please try again." and let the student try the same question repeatedly until the student gets it right. A separate method should be used to generate each new question. This method should be called once when the app begins execution and each time the user answers the question correctly.