BarChart
This program is to accept the number of points scored by each player in a season. The bar chart displays one asterisk for each 10 points scored by a player. For example, if a player has scored 48 points, then display 4 asterisks.
An example of the program is shown below:
Please enter an integer between 1 and 30:
3
Please enter an integer between 1 and 30:
5
Please enter an integer between 1 and 30:
8
Please enter an integer between 1 and 30:
13
Please enter an integer between 1 and 30:
21
***
*****
********
*************
*********************